Ajna library wrapper
